+++ paascloud-provider/paascloud-provider-omc/src/main/java/com/paascloud/provider/service/OmcShippingService.java
@@ -42,7 +42,7 @@ public interface OmcShippingService extends IService<OmcShipping> {
 	 *
 	 * @return the int
 	 */
-	int deleteShipping(Long userId, Integer shippingId);
+	int deleteShipping(Long userId, Long shippingId);
 
 	/**
 	 * 根据Id查询收货人地址.
